코드 중간 리팩토링 필요없는 파일 삭제 필요없는 컴포넌트 삭제 파일명 케밥 케이스로 변경 들여쓰기 2로 설정
