+++ src/app/email-feedback/page.tsx
@@ -1,6 +1,6 @@
 "use client"
 import React, { useState, useRef } from 'react';
-import EmailFeedBack from '@/components/forms/emailFeedBack';
+import EmailFeedBack from '@/components/forms/email-feedback';
 import Image from 'next/image';
 
 const EmailFeedBackPage: React.FC = () => {
