+++ src/app/home/page.tsx
@@ -1,7 +1,9 @@
-import Link from 'next/link';
 import HomeDesign from '@/components/home/home';
-export default function Page() {
+
+const HomePage = () => {
     return (<div className="flex justify-center items-center">
         <HomeDesign/>
     </div>);
 }
+
+export default HomePage;
