+++ src/app/page.tsx
@@ -1,4 +1,3 @@
-import Link from 'next/link';
 import HomeDesign from '@/components/home/home';
 
 export default function Page() {
