--- src/components/background/background.css
@@ -1,5 +1,0 @@
-/* src/index.css */
-
-@import 'tailwindcss/base';
-@import 'tailwindcss/components';
-@import 'tailwindcss/utilities';
