--- src/components/buttons/imageButtons/smallImageButton.tsx
@@ -1,33 +1,0 @@
-import Image from 'next/image';
-import React from 'react';
-
-interface SmallImageButtonProps {
-  onClick: () => void;
-  imageSrc: string;
-  alt: string;
-}
-
-const SmallImageButton: React.FC<SmallImageButtonProps> = ({ onClick, imageSrc, alt }) => {
-    return (
-      <button 
-        onClick={onClick} 
-        className="
-          inline-block relative
-          p-1 border-2 border-purple-700 rounded-md
-          sm:p-2 sm:border-3 sm:rounded-lg
-          md:p-3 md:border-4 md:rounded-xl
-          lg:p-4 lg:border-5 lg:rounded-2xl
-        "
-      >
-        <Image 
-          src={imageSrc} 
-          alt={alt} 
-          width={48} // Adjust as needed
-          height={48} // Adjust as needed
-          className="rounded-md sm:rounded-lg md:rounded-xl lg:rounded-2xl"
-        />
-      </button>
-    );
-  };
-  
-export default SmallImageButton;
