--- src/components/objects/cards/customCards/card.css
@@ -1,68 +1,0 @@
-/* card.css */
-
-.card-container {
-  /* 기존 스타일 유지 */
-  max-width: 20rem;
-  border-radius: 1rem;
-  overflow: hidden;
-  box-shadow: 0 0.5rem 1rem rgba(0, 0, 0, 0.1);
-  transform-style: preserve-3d;
-  transition: transform 0.3s ease;
-  cursor: pointer;
-  position: relative;
-
-  /* 유리 재질 효과를 위한 스타일 */
-  background: rgba(255, 255, 255, 0.3); /* 투명한 배경 */
-  backdrop-filter: blur(10px); /* 배경 블러 효과 */
-  border: 1px solid rgba(255, 255, 255, 0.3); /* 투명한 테두리 */
-  box-shadow: 0 4px 6px rgba(0, 0, 0, 0.1), 0 1px 3px rgba(0, 0, 0, 0.08); /* 그림자 추가 */
-}
-
-/* 나머지 스타일 유지 */
-
-
-  .card-container:hover {
-    transform: translate(-0.25rem, -0.25rem);
-  }
-
-  .card-tilt-up {
-    transform: rotateX(-30deg);
-  }
-
-  .card-tilt-down {
-    transform: rotateX(30deg);
-  }
-
-  .card-tilt-left {
-    transform: rotateY(-30deg);
-  }
-
-  .card-tilt-right {
-    transform: rotateY(30deg);
-  }
-
-  .zoomed {
-    transform: perspective(600px) rotateX(0deg) rotateY(0deg) scale(2);
-  }
-
-  .clicked {
-    transform: perspective(600px) rotateX(0deg) rotateY(360deg) scale(2);
-  }
-
-  /* Shadow Effect */
-  .card-shadow {
-    position: absolute;
-    top: 0;
-    left: 0;
-    width: 100%;
-    height: 100%;
-    background-color: rgba(0, 0, 0, 0.1); /* Adjust the shadow color */
-    transform-origin: center;
-    transform: scale(0.9); /* Adjust the shadow scale */
-    pointer-events: none; /* Prevent shadow from blocking mouse events */
-    transition: transform 0.7s ease;
-  }
-
-  .card-container:hover .card-shadow {
-    transform: scale(1);
-  }
