--- src/components/objects/cards/customCards/clusterCard.css
@@ -1,15 +1,0 @@
-.card {
-    width: 400px;
-    height: 300px;
-    background-color: #3498db;
-    transform: perspective(900px) rotateY(20deg);
-    transition: transform 0.3s ease;
-    cursor: pointer;
-    box-shadow: 0px 8px 12px rgba(0, 0, 0, 0.2); /* 그림자 스타일 설정 */
-    border: solid black 1px;
-  }
-  
-  .card.rotated {
-    transform: perspective(900px) rotateY(0deg) 0.5s ease;
-  }
-
